backend/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
PyJWT==2.1.0
six==1.15.0
SQLAlchemy==1.4.11
Werkzeug==1.0.1
